feat: 주문/결제 api 구현

+++ apps/commerce-api/src/main/kotlin/com/loopers/domain/point/ChargePointCommand.kt
@@ -4,5 +4,5 @@ import com.loopers.domain.user.LoginId
 
 data class ChargePointCommand(
     val loginId: LoginId,
-    val point: Int,
+    val point: Long,
 )
